API documentation

Exported structures and functions

CitableCollection.PropertyDefinition — Type

Configuration data for a single property in a collection.

CitableCollection.citetype — Function

Retrieve CITE type for property.

citetype(pdef)

CitableCollection.authlist — Function

Retrieve (possibly empty) list of allowed values.

authlist(pdef)

CitableCollection.RawDataCollection — Type

A collection of citable data.

The collection is itself a citable object. In addition, it contains a table of citable objects represented as a TypedTables.Table.

Internal functions

CitableCollection.lazyread — Function

Read CITE collections from a CEX source without reference to citeproperties configuration.

lazyread(cexsrc)
lazyread(cexsrc, delimiter)

Collections are assumed to have a field named urn that has the unique identifier for objects in the collection.

CitableCollection.strictread — Function

Read data for citable collections from cexsrc. Determine types of columns for each table from configuration in citeproperties section of the CEX.

strictread(cexsrc)
strictread(cexsrc, delimiter)

If any columns of a collection are not configured, it is an error.

CitableCollection.collectionsfromprops — Function

Compute a list of URNs identifying collections from a list of PropertyDefinitions.

collectionsfromprops(proplist)

CitableCollection.propertyids — Function

Compute property names for properties in a list of PropertyDefinitions matching a give collection URN.

propertyids(propslist, collectionurn)

CitableCollection.typeforcprop — Function

Determine Julia type corresponding to CEX name for property type.

typeforcprop(cprop)

CitableCollection.cpropfortype — Function

Find CITE string value to use in serializing property type.

cpropfortype(pd)

Find CITE string value to use in serializing property type.

cpropfortype(citeproptype)

CitableCollection.converttypes — Function

Convert any string-value columns in a RawDataCollection configured for URN values in their property definitions to the appropriate type of URN value.

converttypes(rdc, rdcprops)