API documentation

Exported structures

CitableCollection.CatalogedCollection — Type

A collection of citable objects.

CitableCollection.PropertyConfiguration — Type

Configuration data for a single property in a collection.

Exported functions

CitableCollection.catalogdf — Function

Parse CEX source data into a catalog of CitableCollections.

catalogdf(cexsrc)
catalogdf(cexsrc, delim)

cexsrc must have at least one citecollections block and one citeproperties block.

CitableCollection.collectiondfs — Function

Use catalog to parse CEX source data into a Vector of DataFrames.

collectiondfs(cexsrc, catalogdf)
collectiondfs(cexsrc, catalogdf, delim; limitto)

cexsrc must have at least one citedata block with data records matching information in catalog.

Internal functions

CitableCollection.blocksagree — Function

True if list of URNs from collections block and properties block document the same set of collections.

blocksagree(collurns, propurns)

CitableCollection.catalogdata — Function

From a list of CEX blocks, extract all data lines from one or more citecollections blocks.

catalogdata(blocklist)

It is a DomainError if blocklist includes no citecollections blocks.

CitableCollection.collectionurns — Function

Extract URNs of collections documented in datalines.

collectionurns(datalines)
collectionurns(datalines, delim)

datalines is a Vector of delimited-string values.

CitableCollection.propertydata — Function

From a list of CEX blocks, extract all data lines from one or more citeproperties blocks.

propertydata(blocklist)

It is a DomainError if blocklist includes no citeproperties blocks.

CitableCollection.propertyurns — Function

Extract collection-level URNs of properties documented in datalines.

propertyurns(datalines)
propertyurns(datalines, delim)

CitableCollection.propertyconfigs — Function

Create a DataFrame of PropertyConfigurations from delimited-text source.

propertyconfigs(datalines)
propertyconfigs(datalines, delim)

CitableCollection.configured_propertynames — Function

Extract list of property names from a DataFrame of PropertyConfiguration objects.

configured_propertynames(df)